refactor: 불필요한 코드 제거
