+++ package.json
@@ -18,7 +18,6 @@
     "recoil": "^0.7.7",
     "styled-components": "^6.1.11",
     "styled-reset": "^4.5.2",
-    "undici": "^6.17.0",
     "vite-plugin-environment": "^1.1.3"
   },
   "devDependencies": {
@@ -40,9 +39,7 @@
     "eslint-plugin-react-hooks": "^4.6.0",
     "eslint-plugin-react-refresh": "^0.4.6",
     "jest-environment-jsdom": "^29.7.0",
-    "msw": "^2.3.0",
     "react-router-dom": "^6.23.1",
-    "recoil-persist": "^5.1.0",
     "ts-jest": "^29.1.2",
     "ts-node": "^10.9.2",
     "typescript": "^5.2.2",
