--- src/constants/button.ts
@@ -1,33 +1,0 @@
-export const BUTTON_THEME = {
-  black: {
-    backgroundColor: "#000000",
-    color: "#ffffff",
-  },
-  white: {
-    backgroundColor: "#ffffff",
-    color: "#000000",
-  },
-  disabled: {
-    backgroundColor: "#BEBEBE",
-    color: "#ffffff",
-  },
-};
-
-export const BUTTON_SIZE = {
-  xs: {
-    width: "24px",
-    height: "24px",
-  },
-  s: {
-    width: "40px",
-    height: "24px",
-  },
-  m: {
-    width: "56px",
-    height: "32px",
-  },
-  full: {
-    width: "100%",
-    height: "100%",
-  },
-};
