+++ src/constants/cart.ts
@@ -1,8 +1,3 @@
-export const API_URL = process.env.API_URL || "";
-export const USER_ID = process.env.USER_ID || "";
-export const USER_PASSWORD = process.env.USER_PASSWORD || "";
-export const ENVIRONMENT = process.env.ENVIRONMENT || "";
-
 export const DEFAULT_DELIVERY_FEE = 3_000;
 export const DELIVERY_FEE_THRESHOLD = 100_000;
