+++ src/constants/coupon.ts
@@ -1,0 +1,1 @@
+export const MAX_COUPON_NUMBER = 2;
