+++ src/constants/environment.ts
@@ -1,0 +1,4 @@
+export const API_URL = process.env.API_URL || "";
+export const USER_ID = process.env.USER_ID || "";
+export const USER_PASSWORD = process.env.USER_PASSWORD || "";
+export const ENVIRONMENT = process.env.ENVIRONMENT || "";
